enable category GET POST PUT PATCH and DELETE operation

+++ src/main/java/net/samitkumar/ecom_product_catalogue/model/category/Category.java
@@ -4,5 +4,4 @@ import net.samitkumar.ecom_product_catalogue.model.product.Product;
 
 import java.util.List;
 
-public record Category(int id, String name, String description, List<Product> products) {
-}
+public record Category(Long id, String name, String description, List<Product> products) { }
